Remove post request

+++ src/api/operations/user_operations.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"github.com/go-chi/render"
+	"github.com/younny/slobbo-backend/src/auth"
 	m "github.com/younny/slobbo-backend/src/middleware"
 	"github.com/younny/slobbo-backend/src/types"
 )
@@ -58,6 +59,17 @@ func (server *Server) UpdateUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	tokenID, err := auth.ExtractTokenID(r)
+	if err != nil {
+		_ = render.Render(w, r, types.ErrNotAuthorised(err))
+		return
+	}
+
+	if tokenID != uint32(user.ID) {
+		_ = render.Render(w, r, types.ErrNotAuthorised(err))
+		return
+	}
+
 	user.Prepare()
 
 	if err := user.Validate(""); err != nil {
@@ -78,6 +90,17 @@ func (server *Server) UpdateUser(w http.ResponseWriter, r *http.Request) {
 
 func (server *Server) DeleteUser(w http.ResponseWriter, r *http.Request) {
 	user := r.Context().Value(m.UserCtxKey).(*types.User)
+
+	tokenID, err := auth.ExtractTokenID(r)
+	if err != nil {
+		_ = render.Render(w, r, types.ErrNotAuthorised(err))
+		return
+	}
+
+	if tokenID != uint32(user.ID) {
+		_ = render.Render(w, r, types.ErrNotAuthorised(err))
+		return
+	}
 
 	if err := server.DB.DeleteUser(user.ID); err != nil {
 		_ = render.Render(w, r, types.ErrInvalidRequst(err))
